Avoid leaking Response mapping outside base module
