+++ golem-worker-service-base/src/lib.rs
@@ -4,7 +4,7 @@ pub mod api;
 pub mod repo;
 pub mod service;
 pub mod auth;
-pub mod worker_request;
+pub mod worker_bridge;
 pub mod metrics;
 pub mod app_config;
 mod definition;
